requirements.txt
asgiref==3.8.1
faker
djangorestframework
celery
django-celery-results
django-debug-toolbar
async-timeout==5.0.1
Django==5.2
psycopg2-binary==2.9.10
redis==6.0.0
sqlparse==0.5.3
typing_extensions==4.13.2
